changed Database, updated controllers

+++ XorHub/XorHub/Batch.cs
@@ -21,7 +21,7 @@ namespace XorHub
             this.LoginInfoes = new HashSet<LoginInfo>();
         }
     
-        public decimal BatchId { get; set; }
+        public int BatchId { get; set; }
         public string Name { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
